Npp now opens txt by default, npp now installs by default. Explorer only restarts once.

+++ New-SandboxTemplate.ps1
@@ -36,8 +36,8 @@ Param(
 	[Parameter(HelpMessage="Do not make basic config changes, like show file extensions.")]
 	[Switch]$NoBasicConfig,
 
-	[Parameter(HelpMessage="Install Notepad++. Default: Do not install.")]
-	[switch]$InstallNotepadPlusPlus
+	[Parameter(HelpMessage="Do not install Notepad++. Default: Install.")]
+	[switch]$DontInstallNotepadPlusPlus
 
 )
 
@@ -52,6 +52,7 @@ BEGIN {
 
 	$Template = "<Configuration>`n"
 	$LogonCommands = @()
+	$NeedExplorerRestart = $false
 
 }
 
@@ -139,7 +140,39 @@ PROCESS {
 			Set-ItemProperty -Path "HKCU:\SOFTWARE\Microsoft\Windows\CurrentVersion\Explorer\Advanced" -Name "LaunchTo" -Value 1 -Type DWord -Force
 			Write-Output "[$(Get-Date)] Set launch-to my computer" | Out-File -FilePath "$HOME\Desktop\install_log.txt" -Append
 		}
-		$ConfigRestartExplorerCommand = {
+		$NeedExplorerRestart = $true
+		$LogonCommands += "`t`t<Command>powershell.exe -ExecutionPolicy Bypass -EncodedCommand " + [System.Convert]::ToBase64String([System.Text.Encoding]::Unicode.GetBytes($ConfigFileextensionsCommand.ToString())) + "</Command>`n"
+		$LogonCommands += "`t`t<Command>powershell.exe -ExecutionPolicy Bypass -EncodedCommand " + [System.Convert]::ToBase64String([System.Text.Encoding]::Unicode.GetBytes($ConfigLaunchtocomputerCommand.ToString())) + "</Command>`n"
+	}
+
+
+
+	### =======================================================================
+	### Installers ------------------------------------------------------------
+	### =======================================================================
+
+	if (-not $DontInstallNotepadPlusPlus) {
+		$NPPCommand = {
+			$Response = Invoke-WebRequest -Uri "https://github.com/notepad-plus-plus/notepad-plus-plus/releases/latest" -UseBasicParsing
+			$Version = $Response.BaseResponse.ResponseUri.AbsoluteUri.Split("/")[-1]
+			$InstallerFilename = "npp."
+			$InstallerFilename += $Version.Replace("v", "")
+			$InstallerFilename += ".Installer.x64.exe"
+			$InstallerURL = "https://github.com/notepad-plus-plus/notepad-plus-plus/releases/download/$Version/$InstallerFilename"
+			Invoke-WebRequest -Uri $InstallerURL -OutFile (Join-Path "$HOME\Downloads" $InstallerFilename)
+			& "$HOME\Downloads\npp.8.8.1.Installer.x64.exe" /S
+			Write-Output "[$(Get-Date)] Installed Notepad++" | Out-File -FilePath "$HOME\Desktop\install_log.txt" -Append
+			cmd /c assoc .txt="npptxt"
+			cmd /c  --% ftype npptxt="C:\Program Files\Notepad++\notepad++.exe" "%1"
+			Write-Output "[$(Get-Date)] Associated .txt with Notepad++" | Out-File -FilePath "$HOME\Desktop\install_log.txt" -Append
+		}
+		$LogonCommands += "`t`t<Command>powershell.exe -ExecutionPolicy Bypass -EncodedCommand " + [System.Convert]::ToBase64String([System.Text.Encoding]::Unicode.GetBytes($NPPCommand.ToString())) + "</Command>`n"
+	}
+
+
+
+	if ($NeedExplorerRestart) {
+		$RestartExplorerCommand = {
 			Stop-Process -Name 'Explorer' -Force
 			Start-Sleep -Seconds 3
 			try {
@@ -155,34 +188,9 @@ PROCESS {
 			}
 			Write-Output "[$(Get-Date)] Restarted explorer" | Out-File -FilePath "$HOME\Desktop\install_log.txt" -Append
 		}
-		$LogonCommands += "`t`t<Command>powershell.exe -ExecutionPolicy Bypass -EncodedCommand " + [System.Convert]::ToBase64String([System.Text.Encoding]::Unicode.GetBytes($ConfigFileextensionsCommand.ToString())) + "</Command>`n"
-		$LogonCommands += "`t`t<Command>powershell.exe -ExecutionPolicy Bypass -EncodedCommand " + [System.Convert]::ToBase64String([System.Text.Encoding]::Unicode.GetBytes($ConfigLaunchtocomputerCommand.ToString())) + "</Command>`n"
-		$LogonCommands += "`t`t<Command>powershell.exe -ExecutionPolicy Bypass -EncodedCommand " + [System.Convert]::ToBase64String([System.Text.Encoding]::Unicode.GetBytes($ConfigRestartExplorerCommand.ToString())) + "</Command>`n"
+		$LogonCommands += "`t`t<Command>powershell.exe -ExecutionPolicy Bypass -EncodedCommand " + [System.Convert]::ToBase64String([System.Text.Encoding]::Unicode.GetBytes($RestartExplorerCommand.ToString())) + "</Command>`n"
 	}
-
-
-
-	### =======================================================================
-	### Installers ------------------------------------------------------------
-	### =======================================================================
-
-	if ($InstallNotepadPlusPlus) {
-		$NPPCommand = {
-			$Response = Invoke-WebRequest -Uri "https://github.com/notepad-plus-plus/notepad-plus-plus/releases/latest" -UseBasicParsing
-			$Version = $Response.BaseResponse.ResponseUri.AbsoluteUri.Split("/")[-1]
-			$InstallerFilename = "npp."
-			$InstallerFilename += $Version.Replace("v", "")
-			$InstallerFilename += ".Installer.x64.exe"
-			$InstallerURL = "https://github.com/notepad-plus-plus/notepad-plus-plus/releases/download/$Version/$InstallerFilename"
-			Invoke-WebRequest -Uri $InstallerURL -OutFile (Join-Path "$HOME\Downloads" $InstallerFilename)
-			& "$HOME\Downloads\npp.8.8.1.Installer.x64.exe" /S
-			Write-Output "[$(Get-Date)] Installed Notepad++" | Out-File -FilePath "$HOME\Desktop\install_log.txt" -Append
-		}
-		$LogonCommands += "`t`t<Command>powershell.exe -ExecutionPolicy Bypass -EncodedCommand " + [System.Convert]::ToBase64String([System.Text.Encoding]::Unicode.GetBytes($NPPCommand.ToString())) + "</Command>`n"
-	}
-
-
-
+	
 	if ($LogonCommands.Count -gt 0) {
 		$Template += "`t<LogonCommand>`n"
 		$Template += $LogonCommands -join ""
